requirements.txt
Django==5.0.2
djangorestframework==3.16.1
django-filter==25.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
pre_commit==4.3.0
python-json-logger==3.3.0
psycopg2-binary==2.9.10
python-keycloak==5.8.1
requests==2.32.5
openpyxl==3.1.5
pandas==2.3.2
graphene-django==3.2.3